+++ demisto_sdk/__main__.py
@@ -12,7 +12,11 @@ import typer
 from pkg_resources import DistributionNotFound, get_distribution
 
 from demisto_sdk.commands.common.configuration import Configuration
-from demisto_sdk.commands.common.constants import ENV_DEMISTO_SDK_MARKETPLACE, FileType
+from demisto_sdk.commands.common.constants import (
+    ENV_DEMISTO_SDK_MARKETPLACE,
+    FileType,
+    MarketplaceVersions,
+)
 from demisto_sdk.commands.common.content_constant_paths import (
     ALL_PACKS_DEPENDENCIES_DEFAULT_PATH,
 )
@@ -913,12 +917,12 @@ def secrets(config, **kwargs):
     type=int,
 )
 @click.option(
-    "-idp",
-    "--id-set-path",
-    help="Path to id_set.json, relevant for when using the "
-    "--check-dependent-api-module flag.",
-    type=click.Path(resolve_path=True),
-    default="Tests/id_set.json",
+    "-di",
+    "--docker-image",
+    default="from-yml",
+    help="The docker image to check package on. Possible values: 'native:maintenance', 'native:ga', 'native:dev',"
+    " 'all', a specific docker image from Docker Hub (e.g devdemisto/python3:3.10.9.12345) or the default"
+    " 'from-yml'.",
 )
 @click.option(
     "-cdam",
@@ -940,7 +944,7 @@ def lint(**kwargs):
     2. Package in docker image checks -  pylint, pytest, powershell - test, powershell - analyze.
     Meant to be used with integrations/scripts that use the folder (package) structure.
     Will lookup up what docker image to use and will setup the dev dependencies and file in the target folder.
-    If no additional flags specifying the packs are given,will lint only changed files.
+    If no additional flags specifying the packs are given, will lint only changed files.
     """
     from demisto_sdk.commands.common.logger import logging_setup
     from demisto_sdk.commands.lint.lint_manager import LintManager
@@ -960,7 +964,6 @@ def lint(**kwargs):
         quiet=kwargs.get("quiet"),  # type: ignore[arg-type]
         prev_ver=kwargs.get("prev_ver"),  # type: ignore[arg-type]
         json_file_path=kwargs.get("json_file"),  # type: ignore[arg-type]
-        id_set_path=kwargs.get("id_set_path"),  # type: ignore[arg-type]
         check_dependent_api_module=kwargs.get("check_dependent_api_module"),  # type: ignore[arg-type]
     )
     return lint_manager.run(
@@ -980,6 +983,7 @@ def lint(**kwargs):
         no_coverage=kwargs.get("no_coverage"),  # type: ignore[arg-type]
         coverage_report=kwargs.get("coverage_report"),  # type: ignore[arg-type]
         docker_timeout=kwargs.get("docker_timeout"),  # type: ignore[arg-type]
+        docker_image_flag=kwargs.get("docker_image"),  # type: ignore[arg-type]
         time_measurements_dir=kwargs.get("time_measurements_dir"),  # type: ignore[arg-type]
     )
 
@@ -2640,6 +2644,13 @@ def test_content(**kwargs):
     "-rn", "--release-notes", is_flag=True, help="Will run only on release notes files"
 )
 @click.option(
+    "-xs",
+    "--xsoar-only",
+    is_flag=True,
+    help="Run only on files from XSOAR-supported Packs.",
+    default=False,
+)
+@click.option(
     "-pkw",
     "--use-packs-known-words",
     is_flag=True,
@@ -2662,6 +2673,7 @@ def doc_review(**kwargs):
         use_git=kwargs.get("use_git"),
         prev_ver=kwargs.get("prev_ver"),
         release_notes_only=kwargs.get("release_notes"),
+        xsoar_only=kwargs.get("xsoar_only"),
         load_known_words_from_pack=kwargs.get("use_packs_known_words"),
     )
     result = doc_reviewer.run_doc_review()
@@ -2917,20 +2929,18 @@ def error_code(config, **kwargs):
 )
 @click.help_option("-h", "--help")
 @click.option(
-    "-ud",
-    "--use-docker",
-    is_flag=True,
-    help="Use docker service to run the content graph",
+    "-o",
+    "--output-path",
+    type=click.Path(resolve_path=True, path_type=Path, dir_okay=True, file_okay=False),
+    default=None,
+    help="Output folder to place the zip file of the graph exported CSVs files",
 )
 @click.option(
-    "-us", "--use-existing", is_flag=True, help="Use existing service", default=False
-)
-@click.option(
-    "-se",
-    "--skip-export",
-    is_flag=True,
-    help="Whether or not to skip exporting to CSV.",
-    default=False,
+    "-mp",
+    "--marketplace",
+    help="The marketplace to generate the graph for.",
+    default="xsoar",
+    type=click.Choice(list(MarketplaceVersions)),
 )
 @click.option(
     "-nd",
@@ -2938,9 +2948,6 @@ def error_code(config, **kwargs):
     is_flag=True,
     help="Whether or not to include dependencies.",
     default=False,
-)
-@click.option(
-    "-o", "--output-file", type=click.Path(), help="dump file output", default=None
 )
 @click.option(
     "-v",
@@ -2961,11 +2968,9 @@ def error_code(config, **kwargs):
     type=click.Path(resolve_path=True),
 )
 def create_content_graph(
-    use_docker: bool = False,
-    use_existing: bool = False,
-    skip_export: bool = False,
+    marketplace: str = MarketplaceVersions.XSOAR,
     no_dependencies: bool = False,
-    output_file: Path = None,
+    output_path: Path = None,
     **kwargs,
 ):
     from demisto_sdk.commands.common.logger import logging_setup
@@ -2978,13 +2983,12 @@ def create_content_graph(
         quiet=kwargs.get("quiet"),  # type: ignore[arg-type]
         log_path=kwargs.get("log_path"),
     )  # type: ignore[arg-type]
-    with Neo4jContentGraphInterface(
-        start_service=not use_existing,
-        output_file=Path(output_file) if output_file else None,
-        use_docker=use_docker,
-    ) as content_graph_interface:
+    with Neo4jContentGraphInterface() as content_graph_interface:
         create_content_graph_command(
-            content_graph_interface, not no_dependencies, export=not skip_export
+            content_graph_interface,
+            marketplace=MarketplaceVersions(marketplace),
+            dependencies=not no_dependencies,
+            output_path=output_path,
         )
 
 
@@ -2994,13 +2998,31 @@ def create_content_graph(
 )
 @click.help_option("-h", "--help")
 @click.option(
-    "-ud",
-    "--use-docker",
-    is_flag=True,
-    help="Use docker service to run the content graph",
+    "-mp",
+    "--marketplace",
+    help="The marketplace the artifacts are created for, that "
+    "determines which artifacts are created for each pack. "
+    "Default is the XSOAR marketplace, that has all of the packs "
+    "artifacts.",
+    default="xsoar",
+    type=click.Choice(list(MarketplaceVersions)),
 )
 @click.option(
-    "-us", "--use-existing", is_flag=True, help="Use existing service", default=False
+    "-g",
+    "--use-git",
+    is_flag=True,
+    show_default=True,
+    default=False,
+    help="Whether to use git to determine the packs to update",
+)
+@click.option(
+    "-i",
+    "--imported-path",
+    type=click.Path(
+        path_type=Path, resolve_path=True, exists=True, file_okay=True, dir_okay=False
+    ),
+    default=None,
+    help="Path to content graph zip file to import",
 )
 @click.option(
     "-p",
@@ -3017,7 +3039,11 @@ def create_content_graph(
     default=False,
 )
 @click.option(
-    "-o", "--output-file", type=click.Path(), help="dump file output", default=None
+    "-o",
+    "--output-path",
+    type=click.Path(resolve_path=True, path_type=Path, dir_okay=True, file_okay=False),
+    default=None,
+    help="Output folder to place the zip file of the graph exported CSVs files",
 )
 @click.option(
     "-v",
@@ -3038,11 +3064,12 @@ def create_content_graph(
     type=click.Path(resolve_path=True),
 )
 def update_content_graph(
-    use_docker: bool = False,
-    use_existing: bool = False,
+    use_git: bool = False,
+    marketplace: MarketplaceVersions = MarketplaceVersions.XSOAR,
+    imported_path: Path = None,
     packs: list = None,
     no_dependencies: bool = False,
-    output_file: Path = None,
+    output_path: Path = None,
     **kwargs,
 ):
     from demisto_sdk.commands.common.logger import logging_setup
@@ -3058,16 +3085,18 @@ def update_content_graph(
         quiet=kwargs.get("quiet"),  # type: ignore[arg-type]
         log_path=kwargs.get("log_path"),
     )  # type: ignore[arg-type]
-
-    with Neo4jContentGraphInterface(
-        start_service=not use_existing,
-        output_file=output_file,
-        use_docker=use_docker,
-    ) as content_graph_interface:
+    if packs and not isinstance(packs, list):
+        # for some reason packs provided as tuple from click interface
+        packs = list(packs)
+    with Neo4jContentGraphInterface() as content_graph_interface:
         update_content_graph_command(
             content_graph_interface,
+            marketplace=MarketplaceVersions(marketplace),
+            use_git=use_git,
+            imported_path=imported_path,
             packs_to_update=packs or [],
             dependencies=not no_dependencies,
+            output_path=output_path,
         )