+++ demisto_sdk/commands/content_graph/tests/update_content_graph_test.py
@@ -2,6 +2,7 @@ import shutil
 from distutils.dir_util import copy_tree
 from pathlib import Path
 from typing import Any, Callable, Dict, List
+from zipfile import ZipFile
 
 import pytest
 
@@ -47,6 +48,7 @@ def setup(mocker):
     )
     mocker.patch.object(neo4j_service, "REPO_PATH", GIT_PATH)
     mocker.patch.object(ContentGraphInterface, "repo_path", GIT_PATH)
+    stop_content_graph()
 
 
 @pytest.fixture
@@ -206,11 +208,10 @@ def _get_pack_by_id(repository: ContentDTO, pack_id: str) -> Pack:
 
 
 def dump_csv_import_files(csv_files_dir: Path) -> None:
-    for is_docker in [True, False]:
-        import_path = neo4j_service.get_neo4j_import_path(is_docker).as_posix()
-        if Path(import_path).is_dir():
-            shutil.rmtree(import_path)
-            copy_tree(csv_files_dir.as_posix(), import_path)
+    import_path = neo4j_service.get_neo4j_import_path().as_posix()
+    if Path(import_path).is_dir():
+        shutil.rmtree(import_path)
+        copy_tree(csv_files_dir.as_posix(), import_path)
 
 
 # COMPARISON HELPER FUNCTIONS
@@ -355,11 +356,6 @@ def _testcase2__pack3__remove_relationship(repository: ContentDTO) -> List[Pack]
 
 
 class TestUpdateContentGraph:
-    @classmethod
-    def teardown_class(cls):
-        """Stops the graph interface service. Runs once, after all tests."""
-        stop_content_graph()
-
     def test_merge_graphs(self):
         """
         Given:
@@ -391,7 +387,7 @@ class TestUpdateContentGraph:
                 )
             )
 
-        with ContentGraphInterface(start_service=True) as interface:
+        with ContentGraphInterface() as interface:
             dump_csv_import_files(
                 TEST_DATA_PATH / "mock_import_files_multiple_repos__valid"
             )
@@ -420,6 +416,7 @@ class TestUpdateContentGraph:
     )
     def test_update_content_graph(
         self,
+        tmp_path,
         repository: ContentDTO,
         commit_func: Callable[[ContentDTO], List[Pack]],
         expected_added_dependencies: List[Dict[str, Any]],
@@ -444,7 +441,7 @@ class TestUpdateContentGraph:
         """
         with ContentGraphInterface() as interface:
             # create the graph with dependencies
-            create_content_graph(interface, export=True, dependencies=True)
+            create_content_graph(interface, dependencies=True, output_path=tmp_path)
             packs_from_graph = interface.search(
                 marketplace=MarketplaceVersions.XSOAR,
                 content_type=ContentType.PACK,
@@ -464,7 +461,10 @@ class TestUpdateContentGraph:
 
             # update the graph accordingly
             update_content_graph(
-                interface, packs_to_update=pack_ids_to_update, dependencies=True
+                interface,
+                packs_to_update=pack_ids_to_update,
+                dependencies=True,
+                output_path=tmp_path,
             )
             packs_from_graph = interface.search(
                 marketplace=MarketplaceVersions.XSOAR,
@@ -477,4 +477,15 @@ class TestUpdateContentGraph:
                 expected_added_dependencies,
                 expected_removed_dependencies,
                 after_update=True,
+            )
+        # make sure that the output file zip is created
+        assert Path.exists(tmp_path / "xsoar.zip")
+        with ZipFile(tmp_path / "xsoar.zip", "r") as zip_obj:
+            zip_obj.extractall(tmp_path / "extracted")
+            # make sure that the extracted files are all .csv
+            extracted_files = list(tmp_path.glob("extracted/*"))
+            assert extracted_files
+            assert all(
+                file.suffix == ".csv" or file.name == "metadata.json"
+                for file in extracted_files
             )